refactor(loadingskeleton): rename .scss files to .module.scss
