+++ packages/loading-skeleton/src/LoadingGraphic/LoadingGraphic.module.scss
