+++ packages/loading-skeleton/src/LoadingGraphic/LoadingGraphic.tsx
@@ -1,8 +1,8 @@
 import React, { HTMLAttributes } from "react"
 import classnames from "classnames"
 import { OverrideClassName } from "@kaizen/component-base"
-import skeletonStyles from "../loading-skeleton.scss"
-import graphicStyles from "./LoadingGraphic.scss"
+import skeletonStyles from "../loading-skeleton.module.scss"
+import graphicStyles from "./LoadingGraphic.module.scss"
 
 export type GraphicSizes =
   | "small"
