+++ packages/loading-skeleton/src/LoadingHeading/LoadingHeading.module.scss
