+++ packages/loading-skeleton/src/LoadingHeading/LoadingHeading.tsx
@@ -2,8 +2,8 @@ import React, { HTMLAttributes } from "react"
 import classnames from "classnames"
 import { OverrideClassName } from "@kaizen/component-base"
 import { HeadingVariants } from "@kaizen/typography"
-import skeletonStyles from "../loading-skeleton.scss"
-import headingStyles from "./LoadingHeading.scss"
+import skeletonStyles from "../loading-skeleton.module.scss"
+import headingStyles from "./LoadingHeading.module.scss"
 
 export interface LoadingHeadingProps
   extends OverrideClassName<HTMLAttributes<HTMLDivElement>> {
