+++ packages/loading-skeleton/src/LoadingInput/LoadingInput.module.scss
