+++ packages/loading-skeleton/src/LoadingParagraph/LoadingParagraph.module.scss
