+++ packages/loading-skeleton/src/LoadingParagraph/LoadingParagraph.tsx
@@ -1,8 +1,8 @@
 import React, { HTMLAttributes } from "react"
 import classnames from "classnames"
 import { OverrideClassName } from "@kaizen/component-base"
-import skeletonStyles from "../loading-skeleton.scss"
-import paragraphStyles from "./LoadingParagraph.scss"
+import skeletonStyles from "../loading-skeleton.module.scss"
+import paragraphStyles from "./LoadingParagraph.module.scss"
 
 export interface LoadingParagraphProps
   extends OverrideClassName<HTMLAttributes<HTMLDivElement>> {
